blueprints api, library created and corresponding urls refactored with api.,lib. prefixes
