+++ homework_06/app/app_api/__init__.py
@@ -1,0 +1,1 @@
+from .routes import bp
