+++ homework_06/app/app_library/__init__.py
@@ -1,0 +1,1 @@
+from .routes import bp
