+++ homework_06/app/templates/library_book.html
@@ -8,7 +8,7 @@
         <a
                 type="button"
                 class="btn btn-outline-light h-25"
-                href="{{ url_for("library") }}"
+                href="{{ url_for("lib.library") }}"
         >
             back&nbspto&nbsplibrary
         </a>
