+++ homework_06/app/templates/macros/book_modal_add.html
@@ -41,7 +41,7 @@
                             <form
                                     method="POST"
                                     action="{{ url_for(
-                                    'book',
+                                    'api.book',
                                     google_book_id=book.google_book_id) }}"
                             >
                                 <div class="form-group">
