+++ homework_06/app/templates/macros/book_modal_delete.html
@@ -46,7 +46,7 @@
                     <a
                             type="button"
                             class="btn btn-outline-danger h-25"
-                            href="{{ url_for("library", del_book=book.google_book_id) }}"
+                            href="{{ url_for("lib.library", del_book=book.google_book_id) }}"
                     >
                         delete&nbspfrom&nbsplibrary
                 </a>
